+++ CHANGELOG.md
@@ -13,12 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Create the `SelectButton` component
 -   `@lumx/core`:
     -   Allow single string additional class argument to BEM `block` and `element` functions
+-   `@lumx/vue`:
+    -   Create the `SelectButton` component
 
 ### Changed
 
 -   `@lumx/react`, `@lumx/vue`:
     -   `SelectTextField`: fallback on the default option name (`getOptionName`) if custom option (slot or render function) does not provide a label
-
 ### Fixed
 
 -   `@lumx/react`, `@lumx/vue`: